Fix: dep cycle by extracting actions in an interface
